Replace older gatherer approach with object-oriented version.
