+++ gather
@@ -11,6 +11,11 @@ import importlib
 
 from utils import utils
 
+options = utils.options_for_gather()
+utils.configure_logging(options)
+utils.mkdir_p(utils.cache_dir())
+utils.mkdir_p(utils.results_dir())
+
 # some metadata about the scan itself
 start_time = utils.local_now()
 start_command = str.join(" ", sys.argv)
@@ -24,15 +29,22 @@ strip_wildcard = re.compile("^(\*.)+")
 strip_redacted = re.compile("^(\?\.)+")
 
 
-def run(options=None, cache_dir="./cache", results_dir="./results"):
+def run(options=None):
 
-    sources = options["gatherers"]
+    if not options["_"]:
+        logging.error("Specify a gatherer.")
+        exit(1)
 
-    suffixes = options.get("suffix")
+    # For now, require a --suffix.
+    suffixes = utils.normalize_suffixes(options.get("suffix"))
+    if suffixes is None:
+        logging.error("--suffix with one or more patterns is required.")
+        exit(1)
+    # Turn the suffix into a regex.
     suffix_pattern = utils.suffix_pattern(suffixes)
 
     # Clear out existing result CSVs, to avoid inconsistent data.
-    for result in glob.glob("%s/*.csv" % results_dir):
+    for result in glob.glob("%s/*.csv" % utils.results_dir()):
         os.remove(result)
 
     # Opt in to include parent (second-level) domains.
@@ -44,7 +56,10 @@ def run(options=None, cache_dir="./cache", results_dir="./results"):
 
     # --parents should be a CSV whose first column is parent domains
     # that will act as a whitelist for which subdomains to gather.
-    parents = get_parent_domains(options, cache_dir=cache_dir)
+    parents = get_parent_domains(options)
+
+    # Import the gatherer(s).
+    sources = options["_"][0].split(",")
 
     # De-duping hostnames. This will cause the system to hold all
     # hostnames in memory at once, but oh well.
@@ -54,21 +69,24 @@ def run(options=None, cache_dir="./cache", results_dir="./results"):
         extra = {}
 
         try:
-            gatherer = importlib.import_module("gatherers.%s" % source)
+            gatherer_module = importlib.import_module(
+                "gatherersabc.%s" % source)
+            gatherer = gatherer_module.Gatherer(suffixes, options)
         except ImportError:
             # If it's not a registered module, allow it to be "hot registered"
             # as long as the user gave us a flag with that name that can be
             # used as the --url option to the URL module.
             if options.get(source):
-                gatherer = importlib.import_module("gatherers.url")
+                gatherer_module = importlib.import_module("gatherersabc.url")
                 extra['name'] = source
+                gatherer = gatherer_module.Gatherer(suffixes, options, extra)
             else:
                 exc_type, exc_value, exc_traceback = sys.exc_info()
                 logging.error("[%s] Gatherer not found, or had an error during loading.\n\tERROR: %s\n\t%s" % (source, exc_type, exc_value))
                 exit(1)
 
         # Iterate over each hostname.
-        for domain in gatherer.gather(suffixes, options, extra):
+        for domain in gatherer.gather():
 
             # Always apply the suffix filter to returned names.
             if not suffix_pattern.search(domain):
@@ -123,7 +141,7 @@ def run(options=None, cache_dir="./cache", results_dir="./results"):
     headers += sources
 
     # Open CSV file.
-    gathered_filename = "%s/%s.csv" % (results_dir, "gathered")
+    gathered_filename = "%s/%s.csv" % (utils.results_dir(), "gathered")
     gathered_file = open(gathered_filename, 'w', newline='')
     gathered_writer = csv.writer(gathered_file)
     gathered_writer.writerow(headers)
@@ -155,11 +173,11 @@ def run(options=None, cache_dir="./cache", results_dir="./results"):
         'end_time': utils.utc_timestamp(end_time),
         'command': start_command
     }
-    utils.write(utils.json_for(metadata), "%s/meta.json" % results_dir)
+    utils.write(utils.json_for(metadata), "%s/meta.json" % utils.results_dir())
 
 
 # Read in parent domains from the first column of a given CSV.
-def get_parent_domains(options, cache_dir="./cache"):
+def get_parent_domains(options):
     parents = options.get("parents")
     if not parents:
         return None
@@ -169,7 +187,7 @@ def get_parent_domains(options, cache_dir="./cache"):
     if parents.startswith("http:") or parents.startswith("https:"):
 
         # Though it's saved in cache/, it will be downloaded every time.
-        parents_path = os.path.join(cache_dir, "parents.csv")
+        parents_path = os.path.join(utils.cache_dir(), "parents.csv")
 
         try:
             response = requests.get(parents)
@@ -192,12 +210,4 @@ def get_parent_domains(options, cache_dir="./cache"):
 
 
 if __name__ == '__main__':
-    options = utils.options_for_gather()
-    utils.configure_logging(options)
-    # If we need to add a way to change these, we can add an option to gather.
-    cache_dir = "./cache"
-    results_dir = "./results"
-    utils.mkdir_p(cache_dir)
-    utils.mkdir_p(results_dir)
-
-    run(options, cache_dir=cache_dir, results_dir=results_dir)
+    run(options)
