+++ gatherers/censys.py
@@ -1,13 +1,14 @@
-import os
-import json
 import csv
+import json
 import logging
+import os
 from typing import List
 
 from google.cloud import bigquery
 from google.oauth2 import service_account
 import google.api_core.exceptions
 
+from gatherersabc.gathererabc import Gatherer
 from utils import utils
 
 # Options:
@@ -36,82 +37,81 @@ from utils import utils
 default_timeout = 60 * 60 * 10
 
 
-def gather(suffixes, options, extra={}):
+class Gatherer(Gatherer):
 
-    # Returns a parsed, processed Google service credentials object.
-    credentials = load_credentials()
+    def gather(self):
 
-    if credentials is None:
-        logging.warn("No BigQuery credentials provided.")
-        logging.warn("Set BIGQUERY_CREDENTIALS or BIGQUERY_CREDENTIALS_PATH environment variables.")
-        exit(1)
+        # Returns a parsed, processed Google service credentials object.
+        credentials = load_credentials()
 
-    # When using this form of instantiation, the client won't pull
-    # the project_id out of the creds, has to be set explicitly.
-    client = bigquery.Client(
-        project=credentials.project_id,
-        credentials=credentials
-    )
+        if credentials is None:
+            logging.warn("No BigQuery credentials provided.")
+            logging.warn("Set BIGQUERY_CREDENTIALS or BIGQUERY_CREDENTIALS_PATH environment variables.")
+            exit(1)
 
-    # Allow override of default timeout (in seconds).
-    timeout = int(options.get("timeout", default_timeout))
+        # When using this form of instantiation, the client won't pull
+        # the project_id out of the creds, has to be set explicitly.
+        client = bigquery.Client(
+            project=credentials.project_id,
+            credentials=credentials
+        )
 
-    # Construct the query.
-    query = query_for(suffixes)
-    logging.debug("Censys query:\n%s\n" % query)
+        # Allow override of default timeout (in seconds).
+        timeout = int(self.options.get("timeout", default_timeout))
 
-    # Hardcode this for now:
-    cache_dir = "./cache"
-    # Plan to store in cache/censys/export.csv.
-    download_path = utils.cache_path("export", "censys", ext="csv",
-                                     cache_dir=cache_dir)
+        # Construct the query.
+        query = query_for(self.suffixes)
+        logging.debug("Censys query:\n%s\n" % query)
 
-    # Reuse of cached data can be turned on with --cache.
-    cache = options.get("cache", False)
-    if (cache is True) and os.path.exists(download_path):
-        logging.warn("Using cached download data.")
+        # Plan to store in cache/censys/export.csv.
+        download_path = utils.cache_path("export", "censys", ext="csv")
 
-    # But by default, fetch new data from the BigQuery API,
-    # and write it to the expected download location.
-    else:
-        logging.warn("Kicking off SQL query job.")
+        # Reuse of cached data can be turned on with --cache.
+        cache = self.options.get("cache", False)
+        if (cache is True) and os.path.exists(download_path):
+            logging.warn("Using cached download data.")
 
-        rows = None
+        # But by default, fetch new data from the BigQuery API,
+        # and write it to the expected download location.
+        else:
+            logging.warn("Kicking off SQL query job.")
 
-        # Actually execute the query.
-        try:
-            # Executes query and loads all results into memory.
-            query_job = client.query(query)
-            iterator = query_job.result(timeout=timeout)
-            rows = list(iterator)
-        except google.api_core.exceptions.Forbidden:
-            logging.warn("Access denied to Censys' BigQuery tables.")
-        except:
-            logging.warn(utils.format_last_exception())
-            logging.warn("Error talking to BigQuery, aborting.")
+            rows = None
 
-        # At this point, the job is complete and we need to download
-        # the resulting CSV URL in results_url.
-        logging.warn("Caching results of SQL query.")
+            # Actually execute the query.
+            try:
+                # Executes query and loads all results into memory.
+                query_job = client.query(query)
+                iterator = query_job.result(timeout=timeout)
+                rows = list(iterator)
+            except google.api_core.exceptions.Forbidden:
+                logging.warn("Access denied to Censys' BigQuery tables.")
+            except:
+                logging.warn(utils.format_last_exception())
+                logging.warn("Error talking to BigQuery, aborting.")
 
-        download_file = open(download_path, 'w', newline='')
-        download_writer = csv.writer(download_file)
-        download_writer.writerow(["Domain"])  # will be skipped on read
+            # At this point, the job is complete and we need to download
+            # the resulting CSV URL in results_url.
+            logging.warn("Caching results of SQL query.")
 
-        # Parse the rows and write them out as they were returned (dupes
-        # and all), to be de-duped by the central gathering script.
-        for row in rows:
-            domains = row['common_name'] + row['dns_names']
-            for domain in domains:
-                download_writer.writerow([domain])
+            download_file = open(download_path, 'w', newline='')
+            download_writer = csv.writer(download_file)
+            download_writer.writerow(["Domain"])  # will be skipped on read
 
-        # End CSV writing.
-        download_file.close()
+            # Parse the rows and write them out as they were returned (dupes
+            # and all), to be de-duped by the central gathering script.
+            for row in rows:
+                domains = row['common_name'] + row['dns_names']
+                for domain in domains:
+                    download_writer.writerow([domain])
 
-    # Whether we downloaded it fresh or not, read from the cached data.
-    for domain in utils.load_domains(download_path):
-        if domain:
-            yield domain
+            # End CSV writing.
+            download_file.close()
+
+        # Whether we downloaded it fresh or not, read from the cached data.
+        for domain in utils.load_domains(download_path):
+            if domain:
+                yield domain
 
 
 # Constructs the query to run in BigQuery, against Censys'
