+++ gatherers/rdns.py
@@ -2,6 +2,10 @@ import json
 import logging
 import re
 
+from typing import Generator, List, Pattern
+
+from gatherersabc.gathererabc import Gatherer
+
 # Reverse DNS
 #
 # Given a path to a (local) "JSON Lines" formatted file,
@@ -24,30 +28,38 @@ ip_filter = re.compile("^(\w+[\-\.]?)?\d+[\-\.]\d+[\-\.]\d+[\-\.]\d+")
 number_filter = re.compile("^[\d\-]+\.")
 
 
-def gather(suffixes, options, extra={}):
-    path = options.get("rdns")
+class Gatherer(Gatherer):
 
-    if path is None:
-        logging.warn("--rdns is required to be a path to a local file.")
-        exit(1)
+    def gather(self):
+        path = self.options.get("rdns")
 
-    # May become useful to allow URLs in future.
-    if path.startswith("http:") or path.startswith("https:"):
-        logging.warn("--rdns is required to be a path to a local file.")
-        exit(1)
+        if path is None:
+            logging.warn("--rdns is required to be a path to a local file.")
+            exit(1)
 
-    with open(path) as lines:
-        logging.debug("\tReading %s..." % path)
+        # May become useful to allow URLs in future.
+        if path.startswith("http:") or path.startswith("https:"):
+            logging.warn("--rdns is required to be a path to a local file.")
+            exit(1)
 
-        for line in lines:
-            record = json.loads(line)
-            # logging.debug("\t%s" % record["value"])
+        with open(path) as lines:
+            logging.debug("\tReading %s..." % path)
 
-            # Filter out IP-like reflected addresses.
-            is_ip = (ip_filter.search(record["value"]) is not None)
+            for record in process_lines(lines, ip_filter, number_filter):
+                yield record
 
-            # Check if it's just something like '1234.what.ever.gov'
-            is_number = (number_filter.search(record["value"]) is not None)
 
-            if (not is_ip) and (not is_number):
-                yield record["value"]
+def process_lines(lines: List[str], ip_filter: Pattern,
+                  number_filter: Pattern) -> Generator[str, str, None]:
+    for line in lines:
+        record = json.loads(line)
+        # logging.debug("\t%s" % record["value"])
+
+        # Filter out IP-like reflected addresses.
+        is_ip = (ip_filter.search(record["value"]) is not None)
+
+        # Check if it's just something like '1234.what.ever.gov'
+        is_number = (number_filter.search(record["value"]) is not None)
+
+        if (not is_ip) and (not is_number):
+            yield record["value"]
